avoid a compiler warning

+++ src/kernels/tpetra/kernels_def.hpp
@@ -786,7 +786,7 @@ extern "C" void SUBR(mvec_dot_mvec)(TYPE(const_mvec_ptr) v,
   PHIST_CAST_PTR_FROM_VOID(const Traits<_ST_>::mvec_t, mvec1, v, *iflag);
   PHIST_CAST_PTR_FROM_VOID(const Traits<_ST_>::mvec_t, mvec2, w, *iflag);
 
-  Teuchos::ArrayView<_ST_> result{s, mvec1->getNumVectors()};
+  Teuchos::ArrayView<_ST_> result{s, (int)mvec1->getNumVectors()};
 
   PHIST_TRY_CATCH(mvec1->dot(*mvec2, result), *iflag);
